refactor: 유틸리티 함수 폴더로 정리 및 새로운 함수 추가
