+++ src/utils/array/match.js
@@ -1,0 +1,4 @@
+export const containsElement = (array, element) => array.includes(element);
+
+export const countArrayMatches = (sourceArray, targetArray) =>
+  sourceArray.filter((element) => targetArray.includes(element)).length;
