+++ src/utils/numbers/calculateTicketCount.js
@@ -1,0 +1,3 @@
+export const calculateTicketCount = (amount, priceUnit) => {
+  return Math.floor(amount / priceUnit);
+};
