+++ src/utils/numbers/getRandomNumber.js
@@ -1,0 +1,3 @@
+export const getRandomNumber = (min, max) => {
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+};
